WA: didnt check memory contraint

+++ Solved Problems/Easy/2525_Categorize_Box_According_to_Criteria.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     string categorizeBox(int length, int width, int height, int mass) {
-        bool bulky = (length >= 10e4 || width >= 10e4 || height >= 10e4 || length * width * height >= 10e9) ? true : false, heavy = mass >= 100 ? true : false;
+        bool bulky = (length >= 10e4 || width >= 10e4 || height >= 10e4 || (long long int) length * width * height >= 10e9) ? true : false, heavy = mass >= 100 ? true : false;
         if(bulky && heavy) return "Both";
         if(!bulky && !heavy) return "Neither";
         if(bulky && !heavy) return "Bulky";
